Ready for your review, Min-Yang
Things to check:
Did I get all CAMS trips?
Decision rule to group trips up from the sub-trip level to trip level
Are the three codes (So SES-SAA is an Access Area trip by a LA vessel,
and SES-SCA is an DAS trip by a LA vessels.
And SES-SCG is a trip by a LGC vessel)
the only ones needed?